refactoring code for PR
