+++ tests/test_cross_validation.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 from src.cross_validate import CrossValidation
 from src.data_reader import load_train_test_df
 from src.solutions.constant_predictor import ConstantPredictorSolution
@@ -15,7 +13,7 @@ def test_cross_validation():
     X, y = train_df[x_columns], train_df.drop(columns=['full_text'])
 
     n_splits = 3
-    cv = CrossValidation(saving_dir=Path('/tmp/sdfjsld'), n_splits=n_splits)
+    cv = CrossValidation(saving_dir='/tmp/sdfjsld', n_splits=n_splits)
     predictor = ConstantPredictorSolution()
     cv_scores = cv.fit(predictor, X, y)
     assert cv_scores.shape == (n_splits + 1, 6)
